+++ src/main/java/info/ata4/bspsrc/modules/texture/TextureBuilder.java
@@ -40,7 +40,7 @@ public class TextureBuilder {
     private DTexData texdata;
 
     // indices
-    private short itexinfo = DTexInfo.TEXINFO_NODE;
+    private int itexinfo = DTexInfo.TEXINFO_NODE;
     private int ibrush = -1;
     private int ibrushside = -1;
 
@@ -299,7 +299,7 @@ public class TextureBuilder {
         this.ibrushside = ibrushside;
     }
 
-    public void setTexinfoIndex(short itexinfo) {
+    public void setTexinfoIndex(int itexinfo) {
         this.itexinfo = itexinfo;
     }
